backend/requirements.txt
Django==2.2.2
django-extensions==2.1.7
django-webpack-loader==0.6.0
djangorestframework==3.9.4
mysqlclient==1.3.14
python-dotenv==0.10.3
Werkzeug==0.15.4
yapf==0.25.0
